Clean up websocket events
