--- api/src/main/java/com/ford/labs/retroquest/websocket/WebsocketEvent.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2022 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ford.labs.retroquest.websocket;
-
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-import static com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility.NONE;
-
-@Getter
-@AllArgsConstructor
-@EqualsAndHashCode
-@ToString
-@JsonAutoDetect(fieldVisibility = NONE, getterVisibility = NONE)
-public abstract class WebsocketEvent {
-    @JsonProperty
-    private final WebsocketEventType type;
-    @JsonProperty
-    private final Object payload;
-
-    public abstract String getRoute();
-}
