+++ api/src/main/java/com/ford/labs/retroquest/websocket/events/WebsocketEndRetroEvent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Ford Motor Company
+ * Copyright (c) 2022. Ford Motor Company
  * All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,20 +15,21 @@
  * limitations under the License.
  */
 
-package com.ford.labs.retroquest.websocket;
+package com.ford.labs.retroquest.websocket.events;
 
-import com.ford.labs.retroquest.column.ColumnTitle;
-import org.junit.jupiter.api.Test;
+import static com.ford.labs.retroquest.websocket.events.WebsocketEventType.UPDATE;
 
-import static com.ford.labs.retroquest.websocket.WebsocketEventType.UPDATE;
-import static org.assertj.core.api.Assertions.assertThat;
+public class WebsocketEndRetroEvent extends WebsocketEvent {
+    private static final String ROUTE_STRING = "/topic/%s/end-retro";
+    private final String teamId;
 
-class WebsocketColumnTitleEventTest {
-
-    @Test
-    public void getRoute_returnsTeamSpecificTopic() {
-        var event = new WebsocketColumnTitleEvent("team-id", UPDATE, new ColumnTitle());
-        assertThat(event.getRoute()).isEqualTo("/topic/team-id/column-titles");
+    public WebsocketEndRetroEvent(String teamId) {
+        super(UPDATE, null);
+        this.teamId = teamId;
     }
 
+    @Override
+    public String getRoute() {
+        return String.format(ROUTE_STRING, teamId);
+    }
 }
