+++ api/src/test/java/com/ford/labs/retroquest/websocket/WebsocketServiceTest.java
@@ -18,6 +18,8 @@
 package com.ford.labs.retroquest.websocket;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ford.labs.retroquest.websocket.events.WebsocketEvent;
+import com.ford.labs.retroquest.websocket.events.WebsocketEventType;
 import org.junit.jupiter.api.Test;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
