+++ api/src/test/java/com/ford/labs/retroquest/websocket/events/WebsocketEndRetroEventTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Ford Motor Company
+ * Copyright (c) 2022. Ford Motor Company
  * All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,23 +15,18 @@
  * limitations under the License.
  */
 
-package com.ford.labs.retroquest.websocket;
+package com.ford.labs.retroquest.websocket.events;
 
-import com.ford.labs.retroquest.thought.Thought;
+import com.ford.labs.retroquest.websocket.events.WebsocketEndRetroEvent;
+import org.junit.jupiter.api.Test;
 
+import static org.assertj.core.api.Assertions.assertThat;
 
-public class WebsocketThoughtEvent extends WebsocketEvent {
+class WebsocketEndRetroEventTest {
 
-    private static final String ROUTE_STRING = "/topic/%s/thoughts";
-    private final String teamId;
-
-    public WebsocketThoughtEvent(String teamId, WebsocketEventType type, Thought payload) {
-        super(type, payload);
-        this.teamId = teamId;
-    }
-
-    @Override
-    public String getRoute() {
-        return String.format(ROUTE_STRING, teamId);
+    @Test
+    public void getRoute_returnsTeamSpecificTopic() {
+        var event = new WebsocketEndRetroEvent("team-id");
+        assertThat(event.getRoute()).isEqualTo("/topic/team-id/end-retro");
     }
 }
